+++ src/api/endpoints/admin_routers/auth.py
@@ -1,0 +1,83 @@
+from typing import Tuple
+
+from fastapi import APIRouter, Depends, HTTPException, Request, status
+from fastapi_users import BaseUserManager, models
+from fastapi_users.authentication import AuthenticationBackend, Authenticator, Strategy
+from fastapi_users.manager import UserManagerDependency
+from fastapi_users.openapi import OpenAPIResponseType
+from fastapi_users.router.common import ErrorCode, ErrorModel
+
+from src.api.schemas.admin import OAuth2PasswordRequestForm
+
+
+def get_auth_router(
+    backend: AuthenticationBackend,
+    get_user_manager: UserManagerDependency[models.UP, models.ID],
+    authenticator: Authenticator,
+    requires_verification: bool = False,
+) -> APIRouter:
+    router = APIRouter()
+    get_current_user_token = authenticator.current_user_token(active=True, verified=requires_verification)
+
+    login_responses: OpenAPIResponseType = {
+        status.HTTP_400_BAD_REQUEST: {
+            "model": ErrorModel,
+            "content": {
+                "application/json": {
+                    "examples": {
+                        ErrorCode.LOGIN_BAD_CREDENTIALS: {
+                            "summary": "Bad credentials or the user is inactive.",
+                            "value": {"detail": ErrorCode.LOGIN_BAD_CREDENTIALS},
+                        },
+                        ErrorCode.LOGIN_USER_NOT_VERIFIED: {
+                            "summary": "The user is not verified.",
+                            "value": {"detail": ErrorCode.LOGIN_USER_NOT_VERIFIED},
+                        },
+                    }
+                }
+            },
+        },
+        **backend.transport.get_openapi_login_responses_success(),
+    }
+
+    @router.post(
+        "/login",
+        name=f"auth:{backend.name}.login",
+        responses=login_responses,
+    )
+    async def login(
+        request: Request,
+        credentials: OAuth2PasswordRequestForm = Depends(),
+        user_manager: BaseUserManager[models.UP, models.ID] = Depends(get_user_manager),
+        strategy: Strategy[models.UP, models.ID] = Depends(backend.get_strategy),
+    ):
+        user = await user_manager.authenticate(credentials)
+
+        if user is None or not user.is_active:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail=ErrorCode.LOGIN_BAD_CREDENTIALS,
+            )
+        if requires_verification and not user.is_verified:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail=ErrorCode.LOGIN_USER_NOT_VERIFIED,
+            )
+        response = await backend.login(strategy, user)
+        await user_manager.on_after_login(user, request, response)
+        return response
+
+    logout_responses: OpenAPIResponseType = {
+        **{status.HTTP_401_UNAUTHORIZED: {"description": "Missing token or inactive user."}},
+        **backend.transport.get_openapi_logout_responses_success(),
+    }
+
+    @router.post("/logout", name=f"auth:{backend.name}.logout", responses=logout_responses)
+    async def logout(
+        user_token: Tuple[models.UP, str] = Depends(get_current_user_token),
+        strategy: Strategy[models.UP, models.ID] = Depends(backend.get_strategy),
+    ):
+        user, token = user_token
+        return await backend.logout(strategy, user, token)
+
+    return router
