+++ src/api/endpoints/analytics.py
@@ -6,13 +6,13 @@ from fastapi import APIRouter, Depends, Query
 from src.api.schemas import ActiveTasks, AllUsersStatistic, Analytic, DBStatus, ReasonCancelingStatistics
 from src.api.services import HealthCheckService
 from src.api.services.analytics import AnalyticsService
-from src.authentication import fastapi_users
+from src.core.admin_auth.fastapi_users import fastapi_admin_users
 from src.core.depends import Container
 from src.settings import settings
 
 analytic_router = APIRouter(
     dependencies=[
-        Depends(fastapi_users.current_user(optional=settings.DEBUG)),
+        Depends(fastapi_admin_users.current_user(optional=settings.DEBUG)),
     ]
 )
